make all realTimeMillisecond Integers to Doubles.

+++ include/vrv/doc.h
@@ -203,9 +203,9 @@ public:
      * Run trough all the layers and fill the timemap file content.
      */
     bool ExportTimemap(std::string &output);
-    void PrepareJsonTimemap(std::string &output, std::map<int, double> &realTimeToScoreTime,
-        std::map<int, std::vector<std::string> > &realTimeToOnElements,
-        std::map<int, std::vector<std::string> > &realTimeToOffElements, std::map<int, int> &realTimeToTempo);
+    void PrepareJsonTimemap(std::string &output, std::map<double, double> &realTimeToScoreTime,
+        std::map<double, std::vector<std::string> > &realTimeToOnElements,
+        std::map<double, std::vector<std::string> > &realTimeToOffElements, std::map<double, int> &realTimeToTempo);
 
     /**
      * Set the initial scoreDef of each page.
